refactor(templates): centralize layout components
