+++ app/admin/posts/editar/[slug]/page.tsx
@@ -4,8 +4,8 @@ import { useEffect, useState } from 'react'
 import { useRouter, useParams } from 'next/navigation'
 import { useAuthContext } from '@/lib/context/AuthContext'
 import PostContentEditor from '../../components/PostContentEditor'
-import Header from '@/app/components/Header'
-import Footer from '@/app/components/Footer'
+import Header from '@/components/templates/Header'
+import Footer from '@/components/templates/Footer'
 import Image from 'next/image'
 import { Clock } from 'lucide-react'
 import { isExternalUrl } from '@/utils/isExternalUrl'
