+++ app/blog/layout.tsx
@@ -1,5 +1,5 @@
 import '../globals.css'
-import LayoutWrapper from '../components/LayoutWrapper'
+import LayoutWrapper from '@/components/templates/LayoutWrapper'
 
 export const metadata = {
   title: 'UMADEUS Blog',
