+++ app/login/page.tsx
@@ -4,7 +4,7 @@ import { Suspense, useEffect, useState } from 'react'
 import { useSearchParams } from 'next/navigation'
 import LoginForm from '../components/LoginForm'
 import SignUpForm from '../components/SignUpForm'
-import LayoutWrapper from '../components/LayoutWrapper'
+import LayoutWrapper from '@/components/templates/LayoutWrapper'
 
 export default function LoginPage() {
   return (
